Exclude moc_predefs.h from sync_qt
The QtIfVehicleFunctions generates it's source and header files on the
fly and puts them into the build folder. Those are picked up by sync_qt
to generate the forward headers and the headers.pri

On the first run everything is fine, but once sync_qt is executed again
it picks up other headers from the build folder which are not supposed
to be included, like moc_predefs.h

This causes a broken headers_clean target and breaks the build.

Change-Id: I9c11fb41a2ed88a2d01bb9f54ae26c9d06035e68
Pick-to: 6.2 6.3 6.4
Reviewed-by: Robert Griebl <robert.griebl@qt.io>