Make Vehicle Property examples compatible with Android screens
Vehicle Property examples were written for native displays where the
screen can fit an arbitrarily large window. This is not the case on
Android devices where the screen space is limited. Two affected examples
are changed to have a scalable UI which fit smaller screens and
different orientations. If the screen is too small, flickable areas are
also used, so the whole UI is visible by dragging. QML binding loops
are evaluated as well.

Fixes: QTBUG-110677
Fixes: QTBUG-110678
Pick-to: 6.5
Change-Id: I5276cfc5068bddf0c380e2af7cbc26d262f8db77
Reviewed-by: Dominik Holland <dominik.holland@qt.io>